+++ rs/rosetta-api/icrc1/rosetta/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added /ready endpoint which indicates whether Rosetta is finished with its initial block synch
+- Added Sync Status to network/status
 
 ## [1.1.0] - 2024-06-13
 ### Fixes